turn into package and redo uv stuff
