+++ src/sis_gan/CNN_Subject_Classification.py
@@ -11,8 +11,8 @@ from sklearn.model_selection import StratifiedShuffleSplit
 from torch import nn
 from torch.utils.data import DataLoader, TensorDataset
 
-from models import EEGCNNSubject, weights_init
-from utils import get_accuracy, load_data, load_label, save_model
+from sis_gan.models import EEGCNNSubject, weights_init
+from sis_gan.utils import get_accuracy, load_data, load_label, save_model
 
 logger = logging.getLogger(__name__)
 device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
