+++ src/sis_gan/CNN_Subject_softmax.py
