+++ src/sis_gan/utils.py
